+++ JoinzjazurePHP/Content/bootstrap.cosmo.css
@@ -1,4 +1,4 @@
-@import url("//fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,700");
+@import url("//fonts.useso.com/css?family=Source+Sans+Pro:300,400,700");
 /*!
  * bootswatch v3.2.0
  * Homepage: http://bootswatch.com
