--- JoinzjazurePHP/RenderScripts.php
@@ -1,11 +1,0 @@
-<!--js scripts-->
-<script src="http://cdn.bootcss.com/jquery/3.0.0/jquery.min.js"></script>
-<script src="Scripts/bootstrap.js"></script>
-<script src="Scripts/bootstrap.min.js"></script>
-<script src="Scripts/jquery.validate.js"></script>
-<script src="Scripts/jquery.validate.unobtrusive.js"></script>
-<!--css files-->
-<link rel="stylesheet" href="Content/bootstrap.css">
-<link rel="stylesheet" href="Content/bootstrap.cosmo.css">
-<link rel="stylesheet" href="Content/site.css">
-
