+++ JoinzjazurePHP/Scripts/Home.js
@@ -1,35 +1,1 @@
 ﻿
-<div class='panel panel-primary'>
-<div class='panel-heading'>
-<h4 class='panel-title'>";
-
-
-
-<input type='checkbox' id='ProgramGroup' name='<?php echo $GDF->attributes()->id; ?>' value='<?php echo $GDF->attributes()->value; ?>' onclick='handleCheck('#ProgramGroup',1)'>
-<a data-toggle='collapse' class='info collapsed' data-parent='#accordion' href='#collapseProgramGroup' aria-expanded='false'>程序与移动应用组</a>
-</h4>
-</div>
-
-<div id='collapseProgramGroup' class='panel-collapse collapse' aria-expanded='false' style='height: 0px;'>
-<div class='panel-body'>
-
-</div>
-</div>
-</div>
-
-
-
-
-
-
-<input type='checkbox' id='ProgramGroup' name='<?php echo $GDF->attributes()->id; ?>' value='<?php echo $GDF->attributes()->value; ?>' onclick='handleCheck('#ProgramGroup',1)'>
-<a data-toggle='collapse' class='info collapsed' data-parent='#accordion' href='#collapseProgramGroup' aria-expanded='false'>程序与移动应用组</a>
-</h4>
-</div>
-
-<div id='collapseProgramGroup' class='panel-collapse collapse' aria-expanded='false' style='height: 0px;'>
-<div class='panel-body'>
-
-</div>
-</div>
-</div>
