+++ JoinzjazurePHP/footer.php
@@ -1,0 +1,5 @@
+<footer>
+    <p>&copy; 2013 - <?php echo date('Y')?> 湛江一中IT社 保留所有权利 PHP重制版 By:Nemesiss</p>
+	<script src="Scripts/questions.js"></script>
+</footer>
+</html>
