+++ JoinzjazurePHP/xmlparser.php
@@ -14,3 +14,4 @@
 $Groups = simplexml_load_file("XMLData/Groups.xml");
 $VerificationCodes = simplexml_load_file("XMLData/VerificationCodes.xml");
 
+
